+++ src/main/java/CloneProject/InstagramClone/InstagramService/entity/Following.java
@@ -1,0 +1,28 @@
+package CloneProject.InstagramClone.InstagramService.entity;
+
+import jakarta.persistence.*;
+import lombok.Getter;
+import org.hibernate.annotations.CreationTimestamp;
+
+import java.util.Date;
+
+@Getter
+@Entity
+@Table(name = "followings")
+public class Following {
+    @Id
+    @Column(name = "following_id")
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @CreationTimestamp
+    private Date followTime;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "member_id")
+    private Member following;
+
+    public void setFollowing(Member following) {
+        this.following = following;
+    }
+}
